Refactoring of selected block tools... continued
